# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product catalog
  - name: site
    description: Site/company settings
  - name: enquiries
    description: Contact enquiries
  - name: demo-requests
    description: Demo booking requests
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /public/products:
    get:
      operationId: listPublicProducts
      tags: [products]
      summary: List enabled products
      description: Returns the enabled product/service catalog
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: "#/components/schemas/Product"
                required:
                  - products
  /public/site:
    get:
      operationId: getPublicSiteSettings
      tags: [site]
      summary: Get public company/site settings
      responses:
        "200":
          description: Site settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  settings:
                    $ref: "#/components/schemas/SiteSettings"
                required:
                  - settings
  /enquiries:
    post:
      operationId: createEnquiry
      tags: [enquiries]
      summary: Submit a contact enquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEnquiryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  id:
                    type: integer
                required:
                  - ok
                  - id
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
  /demo-requests:
    post:
      operationId: createDemoRequest
      tags: [demo-requests]
      summary: Submit a demo booking request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDemoRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  id:
                    type: integer
                required:
                  - ok
                  - id
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "409":
          description: Slot no longer available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
  /demo-requests/available-slots:
    get:
      operationId: getDemoRequestAvailableSlots
      tags: [demo-requests]
      summary: List bookable demo time slots for a given date
      parameters:
        - name: date
          in: query
          required: true
          description: Calendar date in YYYY-MM-DD format
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemoAvailableSlots"
        "400":
          description: Invalid date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Product:
      type: object
      properties:
        id:
          type: integer
        key:
          type: string
        title:
          type: string
        category:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        features:
          type: array
          items:
            type: string
        liveUrl:
          type: string
          nullable: true
        enabled:
          type: boolean
        sortOrder:
          type: integer
      required:
        - id
        - key
        - title
        - features
        - enabled
        - sortOrder
    SiteSettings:
      type: object
      properties:
        id:
          type: integer
        companyName:
          type: string
          nullable: true
        tagline:
          type: string
          nullable: true
        primaryEmail:
          type: string
          nullable: true
        supportEmail:
          type: string
          nullable: true
        primaryPhone:
          type: string
          nullable: true
        secondaryPhone:
          type: string
          nullable: true
        addressLine1:
          type: string
          nullable: true
        addressLine2:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        postalCode:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        mapUrl:
          type: string
          nullable: true
      required:
        - id
    CreateEnquiryInput:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        company:
          type: string
        message:
          type: string
      required:
        - firstName
        - lastName
        - email
        - message
    CreateDemoRequestInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        company:
          type: string
        productInterest:
          type: string
        scheduledAt:
          type: string
          description: ISO 8601 date-time of the chosen slot
        message:
          type: string
      required:
        - name
        - email
        - scheduledAt
    DemoSlot:
      type: object
      properties:
        hour:
          type: integer
        label:
          type: string
        iso:
          type: string
        available:
          type: boolean
      required:
        - hour
        - label
        - iso
        - available
    DemoAvailableSlots:
      type: object
      properties:
        date:
          type: string
        timezone:
          type: string
        slots:
          type: array
          items:
            $ref: "#/components/schemas/DemoSlot"
      required:
        - date
        - timezone
        - slots
    ValidationError:
      type: object
      properties:
        error:
          type: string
        issues:
          type: object
      required:
        - error