# lib/api-zod/src/generated/types/createDemoRequestInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateDemoRequestInput {
  name: string;
  email: string;
  phone?: string;
  company?: string;
  productInterest?: string;
  /** ISO 8601 date-time of the chosen slot */
  scheduledAt: string;
  message?: string;
}