# lib/api-zod/src/generated/types/demoAvailableSlots.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DemoSlot } from "./demoSlot";

export interface DemoAvailableSlots {
  date: string;
  timezone: string;
  slots: DemoSlot[];
}