# lib/api-zod/src/generated/types/demoSlot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DemoSlot {
  hour: number;
  label: string;
  iso: string;
  available: boolean;
}